Move EncapsulateFieldInspectionTests for QuickFixes into relevant files under RD.QuickFixes
